fix(ui5-shellbar): fix search field position
Previously the search field has been positioned absolutely and upon resize
it used to remain at the same place, misplaced as all the rest of content is moving.
Now the search field is part of the actions container, moving together with the actions.
Being part of the actions also means we calculate its width to determine how many icons should overflow.

FIXES: https://github.com/SAP/ui5-webcomponents/issues/1570